print_agent/requirements_dashboard.txt
# Requirements untuk Dashboard & Build Executable

# Dashboard (Web-based monitoring)
flask==3.0.0
flask-cors==4.0.0
psutil==5.9.6

# Print agent dependencies (already in requirements.txt)
python-escpos>=3.0
pywin32>=306
requests>=2.31.0

# Build executable
pyinstaller>=6.3.0

# Optional: Better logging
colorama>=0.4.6
